Move product model to models folder
